+++ src/components/Text/Text.test.tsx
@@ -56,34 +56,6 @@ describe('Text', () => {
     expect(screen.getByTestId(TEST_IDS.text.content)).toHaveTextContent('Test default content');
   });
 
-  it('Should apply styles', async () => {
-    const styles = `
-      color: red;
-    `;
-    const replaceVariables = jest.fn((str: string) => str);
-    const props: Props = {
-      data: {} as any,
-      options: {
-        ...DEFAULT_OPTIONS,
-        content: 'Test content',
-        defaultContent: 'Test default content',
-        renderMode: RenderMode.EVERY_ROW,
-        styles,
-      },
-      timeRange: {} as any,
-      timeZone: '',
-      replaceVariables,
-      eventBus: {} as any,
-    };
-
-    render(<Text {...props} />);
-
-    expect(replaceVariables).toHaveBeenCalledWith(styles);
-
-    expect(screen.getByTestId(TEST_IDS.text.content)).toBeInTheDocument();
-    expect(screen.getByTestId(TEST_IDS.text.content)).toHaveStyle({ color: 'red' });
-  });
-
   describe('After Render Function', () => {
     it('Should run after render function', async () => {
       const eventBus = {
@@ -243,6 +215,7 @@ describe('Text', () => {
         },
       ],
     });
+
     const props: Props = {
       data: {} as any,
       frame: dataFrame,
@@ -264,6 +237,54 @@ describe('Text', () => {
     const statuses = screen.getAllByTestId('status');
     expect(statuses[0]).toHaveStyle({ backgroundColor: 'green' });
     expect(statuses[1]).toHaveStyle({ backgroundColor: 'red' });
+  });
+
+  it('Should apply status in row', async () => {
+    const replaceVariables = jest.fn((str: string) => str);
+    const dataFrame = toDataFrame({
+      fields: [
+        {
+          name: 'number',
+          type: FieldType.number,
+          display: (value: number) => ({ color: value > 80 ? 'red' : 'green' }),
+          values: [90],
+        },
+        {
+          name: 'test',
+          type: FieldType.number,
+          display: (value: number) => ({ color: value > 80 ? 'red' : 'green' }),
+          values: [70],
+        },
+        {
+          name: 'value',
+          type: FieldType.number,
+          display: (value: number) => ({ color: value > 80 ? 'red' : 'green' }),
+          values: [60],
+        },
+      ],
+    });
+
+    const props: Props = {
+      data: {} as any,
+      frame: dataFrame,
+      options: {
+        ...DEFAULT_OPTIONS,
+        status: 'number',
+        content: '<div style="background-color: {{statusColor}};" data-testid="status">{{test}}{{number}}</div>',
+        defaultContent: 'Test default content',
+        renderMode: RenderMode.EVERY_ROW,
+      },
+      timeRange: {} as any,
+      timeZone: '',
+      replaceVariables,
+      eventBus: {} as any,
+    };
+
+    render(<Text {...props} />);
+
+    const statuses = screen.getAllByTestId('status');
+
+    expect(statuses[0]).toHaveStyle({ backgroundColor: 'red' });
   });
 
   /**
@@ -465,6 +486,7 @@ describe('Text', () => {
         length: 2,
       }),
     ];
+
     const props: Props = {
       data: {
         series: frames,