+++ CHANGELOG.md
@@ -1,12 +1,24 @@
 # Changelog
 
-## 4.4.0 (IN PROGRESS)
+## 4.4.0 (2024-03-06)
+
+### Breaking changes
+
+- Requires Grafana 9.2 and Grafana 10
 
 ### Features / Enhancements
 
-- Update context (#270)
+- Update context parameter (#270)
 - Added theme object, notifySuccess & notifyError (#270)
 - Update dependencies and Actions (#271)
+- Replace custom code parameters with Code Parameters Builder (#285)
+- Update CSS class for the Panel instead of a Row (#272)
+- Update Editor auto height from fixed value (#278)
+- Disable unitScale and fieldMinMax standard panel options (#286)
+
+### Bug fixes
+
+- Fix statusColor for Row (#280)
 
 ## 4.3.0 (2023-12-25)